Add Milestone 5
- Inserted whole User / Assistant conversation history in function_call
AI context so it has the entire context to provide the most accurate
function call along with the required parameters. The context from both sides
is required because sometimes, the Assistant could make decisions like
choosing a random movie, so that information would be required when the
user proceeds with saying something like, "Buy a ticket now!" without providing
the movie, showtime, etc.